ArC: validate presence of a matching producer for disposer methods
The spec says that it is a definition error if there is no producer method
or producer field declared by the bean class that is assignable to the disposed
parameter of a disposer method. When resolving disposer methods, it is easy
for us to track unused ones, and those are candidates for the validation above.
If a disposer method is used, clearly a matching producer exists.